refactor: 폴더 구조 리팩토링 및 import 경로 수정
- 유틸 관련 파일들을 역할별로 디렉토리 분리
