+++ src/lib/createVNode.js
@@ -1,5 +1,5 @@
-import { flattenChildren } from "../utils/flattenChildren";
-import { removeFalsyChildren } from "../utils/normalizeChildren";
+import { flattenChildren } from "../utils/createVNode/flattenChildren";
+import { removeFalsyChildren } from "../utils/createVNode/normalizeChildren";
 
 export function createVNode(type, props, ...children) {
   const flat = flattenChildren(children);
