+++ src/utils/createVNode/flattenChildren.js
