+++ src/utils/createVNode/normalizeChildren.js
